updates location of db conig and seeding code
