--- backend/src/services/db.connection.js
@@ -1,10 +1,0 @@
-const { Model, knex } = require("objection");
-const Knex = require("knex");
-
-const db = Knex({
-  client: "sqlite3",
-  connection: { filename: "database.sqlite" },
-  useNullAsDefault: true
-});
-
-Model.knex(db);
